add VM
